+++ src/lib/services/types/nft-collection.ts
@@ -2,6 +2,7 @@ import type { Addr, Nullable } from "lib/types";
 
 import {
   zAddr,
+  zBechAddr,
   zHexAddr,
   zHexAddr32,
   zPagination,
@@ -13,15 +14,15 @@ import { z } from "zod";
 
 const zCollection = z
   .object({
-    creator: zHexAddr32,
+    creator: zAddr,
     description: z.string(),
-    id: zHexAddr32,
+    id: zAddr,
     name: z.string(),
     uri: z.string(),
   })
-  .transform((val) => ({
-    ...val,
-    collectionAddress: val.id,
+  .transform(({ id, ...rest }) => ({
+    ...rest,
+    collectionAddress: id,
   }));
 export type Collection = z.infer<typeof zCollection>;
 
@@ -32,7 +33,6 @@ export const zNftCollectionsResponse = z.object({
 export type NftCollectionsResponse = z.infer<typeof zNftCollectionsResponse>;
 
 export interface CollectionByCollectionAddressResponse {
-  createdHeight: Nullable<number>;
   creatorAddress?: Addr;
   currentSupply?: number;
   description: string;
@@ -56,7 +56,6 @@ export const zCollectionByCollectionAddressResponse = z
     }
 
     return {
-      createdHeight: null,
       creatorAddress: undefined,
       currentSupply: undefined,
       description: val.collection.description,
@@ -151,13 +150,17 @@ export type CollectionsByAccountAddressResponse = z.infer<
   typeof zCollectionsByAccountAddressResponse
 >;
 
+// #####################################
+// ############# SEQUENCER #############
+// #####################################
+
 const zCollectionNftsSequencer = z.object({
   handle: zHexAddr32,
   length: z.coerce.number(),
 });
 
 const zCollectionSequencer = z.object({
-  // Revisit this address type
+  // Revisit this address type, it can be empty string
   creator: zHexAddr32,
   description: z.string(),
   name: z.string(),
@@ -171,6 +174,20 @@ export const zCollectionResponseSequencer = z
     object_addr: zHexAddr32,
   })
   .transform(snakeToCamel);
+
+const zTransformedCollectionSequencer =
+  zCollectionResponseSequencer.transform<Collection>((val) => ({
+    collectionAddress: val.objectAddr,
+    ...val.collection,
+  }));
+
+export const zCollectionsResponseSequencer = z.object({
+  collections: z.array(zTransformedCollectionSequencer),
+  pagination: zPagination,
+});
+export type CollectionsResponseSequencer = z.infer<
+  typeof zCollectionsResponseSequencer
+>;
 
 export const zCollectionsByAccountAddressResponseSequencer = z
   .object({
@@ -189,15 +206,25 @@ export type CollectionsByAccountAddressResponseSequencer = z.infer<
   typeof zCollectionsByAccountAddressResponseSequencer
 >;
 
-export const zCollectionByCollectionAddressResponseSequencer = z
+export const zCollectionByCollectionAddressResponseSequencer = z.object({
+  collection: zTransformedCollectionSequencer,
+});
+
+export const zCollectionByCollectionAddressResponseWasm = z
   .object({
-    collection: zCollectionResponseSequencer,
+    data: z.object({
+      creator: zBechAddr,
+      description: z.string(),
+      image: z.string(),
+      royalty_info: z
+        .object({
+          payment_address: zBechAddr,
+          share: z.string(),
+        })
+        .nullable(),
+      start_trading_time: z.string(),
+    }),
   })
-  .transform<CollectionByCollectionAddressResponse>((val) => ({
-    createdHeight: null,
-    creatorAddress: val.collection.collection.creator,
-    currentSupply: val.collection.collection.nfts.length,
-    description: val.collection.collection.description,
-    name: val.collection.collection.name,
-    uri: val.collection.collection.uri,
+  .transform((val) => ({
+    ...snakeToCamel(val.data),
   }));